services/vision-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
paddlepaddle==2.6.2
paddleocr==2.7.3
opencv-python-headless==4.9.0.80
Pillow==10.3.0
httpx==0.27.0
numpy==1.26.4
